Gemmimprove

+++ nd4j-api/src/main/java/org/nd4j/linalg/api/buffer/DoubleBuffer.java
@@ -86,7 +86,7 @@ public class DoubleBuffer extends BaseDataBuffer {
 
     @Override
     public Number getNumber(int i) {
-        return (int) getDouble(i);
+        return  getDouble(i);
     }
 
     @Override
